workspace/revvel_email_organizer/revvel_email_organizer/email_processor.py
"""
Email Archive Processing Module

Handles processing and organization of email archives using AI models.
"""

import os
import logging
from typing import List, Dict, Optional
import yaml

# Configure logging
logging.basicConfig(level=logging.INFO, 
                    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class EmailProcessor:
    """
    Processes and organizes email archives with advanced AI-powered features.
    
    Supports multiple input formats and provides configurable processing strategies.
    """
    
    def __init__(self, archive_path: Optional[str] = None):
        """
        Initialize EmailProcessor with optional custom archive path.
        
        Args:
            archive_path (str, optional): Path to email archives. 
                Defaults to a standard location.
        """
        # Determine default archive location
        if archive_path is None:
            archive_path = os.path.expanduser("~/Emails/Archives")
        
        self.archive_path = archive_path
        os.makedirs(self.archive_path, exist_ok=True)
    
    def process_archive(self, 
                        source_path: str, 
                        model_name: str = 'default', 
                        dry_run: bool = False) -> Dict[str, int]:
        """
        Process an email archive with configurable AI model.
        
        Args:
            source_path (str): Path to the email archive to process
            model_name (str, optional): Name of AI model to use for processing
            dry_run (bool, optional): If True, simulate processing without changes
        
        Returns:
            Dict containing processing statistics
        
        Raises:
            FileNotFoundError: If source archive doesn't exist
            PermissionError: If unable to read or write to archive
        """
        # Validate source path
        if not os.path.exists(source_path):
            raise FileNotFoundError(f"Email archive not found: {source_path}")
        
        try:
            # Simulate processing and collect metrics
            stats = {
                "total_emails": 0,
                "processed_emails": 0,
                "categorized_emails": 0,
                "errors": 0
            }
            
            # Placeholder for actual email processing logic
            # In a real implementation, this would:
            # 1. Parse emails from various formats (mbox, pst, etc.)
            # 2. Apply AI-based categorization
            # 3. Move/tag emails based on content
            
            logger.info(f"Processing email archive: {source_path}")
            logger.info(f"Using model: {model_name}")
            logger.info(f"Dry run mode: {dry_run}")
            
            # Simulated processing (replace with actual implementation)
            stats["total_emails"] = 1000
            stats["processed_emails"] = 950
            stats["categorized_emails"] = 900
            
            if not dry_run:
                self._save_processing_report(stats, source_path)
            
            return stats
        
        except Exception as e:
            logger.error(f"Email processing failed: {e}")
            raise
    
    def _save_processing_report(self, 
                                 stats: Dict[str, int], 
                                 source_path: str):
        """
        Save detailed processing report to archive directory.
        
        Args:
            stats (Dict): Processing statistics
            source_path (str): Original archive path
        """
        try:
            report_filename = os.path.join(
                self.archive_path, 
                f"email_processing_report_{os.path.basename(source_path)}.yaml"
            )
            
            with open(report_filename, 'w') as report_file:
                yaml.safe_dump({
                    "source": source_path,
                    "processed_at": os.path.getctime(source_path),
                    "stats": stats
                }, report_file)
            
            logger.info(f"Processing report saved: {report_filename}")
        
        except Exception as e:
            logger.warning(f"Could not save processing report: {e}")